requirements.txt
fastapi>=0.109.1
uvicorn>=0.27.1
pydantic
streamlit>=1.31.1
pandas>=2.2.0
numpy>=1.26.4
matplotlib
scikit-learn
nltk
spacy
transformers
sqlalchemy>=1.4.0
plotly>=5.18.0
yfinance>=0.2.36
requests>=2.31.0
langchain>=0.1.9
python-dotenv>=1.0.1
google-generativeai>=0.3.2
pandas-ta>=0.3.14b0
langchain-community
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
